Makefile: make all means all
Currently, the `make all` target is not building all: in particular,
it is not building texproc, which means that a `make all` on a clean
tree will fail unless it is preceded by a `make texproc`.

This patch fixes this by including texproc as one of the targets for
`make all`.

Closes: #1